+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/Old/FlywheelAutoAdjust.java
@@ -1,23 +1,23 @@
-package org.firstinspires.ftc.teamcode.Old;
+package org.firstinspires.ftc.robotcontroller.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp; // changed to TeleOp
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.VoltageSensor;
+import com.qualcomm.robotcore.util.ElapsedTime;
 
-
-//Made with AI
 @TeleOp
-public class RtAndLtVelcotyTest extends LinearOpMode {
+public class FlywheelAutoAdjust extends LinearOpMode {
 
     private DcMotorEx flywheel;
     private DcMotor feedRoller;
     private CRServo agitator;
     double flywheelTarget = 900; // starting velocity
     double multiplier;
+    ElapsedTime timer = new ElapsedTime();
 
     public double getLowestVoltage() {
         double lowestValue = Double.POSITIVE_INFINITY;
@@ -34,6 +34,7 @@ public class RtAndLtVelcotyTest extends LinearOpMode {
 
     @Override
     public void runOpMode() {
+
         flywheel = hardwareMap.get(DcMotorEx.class, "flywheel");
         feedRoller = hardwareMap.get(DcMotor.class, "coreHex");
         agitator = hardwareMap.get(CRServo.class, "servo");
@@ -46,20 +47,56 @@ public class RtAndLtVelcotyTest extends LinearOpMode {
 
         waitForStart();
 
+        boolean sequenceStarted = false;
+
         while (opModeIsActive()) {
-            // adjust flywheel target using triggers
+
+            // --- Adjust flywheel velocity with triggers ---
             if (gamepad1.right_trigger > 0.1) {
-                flywheelTarget += 5; // increase
+                flywheelTarget += 5; // increase target
             }
             if (gamepad1.left_trigger > 0.1) {
-                flywheelTarget -= 5; // decrease
+                flywheelTarget -= 5; // decrease target
             }
 
-            // apply voltage compensation
             multiplier = 14 / getLowestVoltage();
             flywheel.setVelocity(flywheelTarget * multiplier);
 
-            // optional telemetry
+            // --- Start the automatic shooting sequence with a button ---
+            if (gamepad1.a && !sequenceStarted) { // press 'A' to start
+                sequenceStarted = true;
+                timer.reset();
+
+                // Wait 1 second for flywheel to spin up
+                while (timer.milliseconds() < 1000 && opModeIsActive()) {
+                    flywheel.setVelocity(flywheelTarget * multiplier);
+                    idle();
+                }
+
+                // Turn on agitator for 0.5 seconds
+                agitator.setPower(1);
+                timer.reset();
+                while (timer.milliseconds() < 500 && opModeIsActive()) {
+                    flywheel.setVelocity(flywheelTarget * multiplier);
+                    idle();
+                }
+
+                // Turn on feed roller for 6 seconds
+                feedRoller.setPower(1);
+                timer.reset();
+                while (timer.milliseconds() < 6000 && opModeIsActive()) {
+                    flywheel.setVelocity(flywheelTarget * multiplier);
+                    agitator.setPower(1);
+                    idle();
+                }
+
+                // Stop everything
+                flywheel.setVelocity(0);
+                agitator.setPower(0);
+                feedRoller.setPower(0);
+            }
+
+            // --- Telemetry ---
             telemetry.addData("Flywheel Target", flywheelTarget);
             telemetry.addData("Flywheel Velocity", flywheel.getVelocity());
             telemetry.addData("Multiplier", multiplier);
