+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/Old/OldTeleOp.java
@@ -1,17 +1,15 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.robotcontroller.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.HardwareDevice;
 import com.qualcomm.robotcore.hardware.VoltageSensor;
 
 
 @TeleOp
-public class FlyWheelteleop extends OpMode {
+public class OldTeleOp extends OpMode {
     private DcMotorEx flywheel;
     private DcMotor feedRoller;
     private DcMotor leftDrive;
@@ -19,12 +17,13 @@ public class FlyWheelteleop extends OpMode {
     private DcMotor rightDrive;
 
     private float flyWheelVelocity = 1300;
-    private float  ticksPerRev = 288;
-    private float offset = 0;
-    // private float FeedRollerSpeed = 65; // No longer needed
+    private static final int bankVelocity = 1300;
+    private static final int farVelocity = 1900;
+    private static final int maxVelocity = 2200;
+
     private boolean flyWheelPowered;
     private boolean agitatorPowered;
-    // private boolean feedRollerPowered; // No longer needed
+    private boolean feedRollerPowered;
 
     public void init() {
         flywheel = hardwareMap.get(DcMotorEx.class, "flywheel");
@@ -33,20 +32,19 @@ public class FlyWheelteleop extends OpMode {
         agitator = hardwareMap.get(CRServo.class, "servo");
         rightDrive = hardwareMap.get(DcMotor.class, "rightDrive");
 
+        // Establishing the direction and mode for the motors
         flywheel.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        feedRoller.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        feedRoller.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
         flywheel.setDirection(DcMotor.Direction.REVERSE);
         feedRoller.setDirection(DcMotor.Direction.REVERSE);
         leftDrive.setDirection(DcMotor.Direction.REVERSE);
         agitator.setDirection(DcMotor.Direction.REVERSE);
         telemetry.addLine("a to turn on/off the flywheel");
         telemetry.addLine("b to turn on/off the agitator");
-        telemetry.addLine("RIGHT/LEFT TRIGGERS to run the feed roller");
-        telemetry.addLine("y to turn off flywheel/agitator and reset feed roller");
+        telemetry.addLine("x to turn on/off the feed roller");
+        telemetry.addLine("Left bumper for slow speed");
+        telemetry.addLine("right bumper for medium speed");
 
+ 
         telemetry.update();
     }
 
@@ -54,9 +52,6 @@ public class FlyWheelteleop extends OpMode {
         basicMovement();
         turnOnMotors();
         flyWheel();
-
-        telemetry.addLine("Encoder Position: " + String.valueOf(feedRoller.getCurrentPosition()));
-        telemetry.addLine("Offset: " + offset);
         telemetry.update();
     }
     public double getLowestVoltage() {
@@ -81,14 +76,10 @@ public class FlyWheelteleop extends OpMode {
         leftDrive.setPower(y - x);
         rightDrive.setPower(y + x);
     }
-
     public void turnOnMotors() {
-        // --- Flywheel (A button) ---
         if(gamepad1.aWasPressed()) {
             flyWheelPowered = !flyWheelPowered;
         }
-
-        // --- Agitator (B button) ---
         if(gamepad1.bWasPressed()) {
             if(agitatorPowered) {
                 agitatorPowered = false;
@@ -98,52 +89,25 @@ public class FlyWheelteleop extends OpMode {
                 agitator.setPower(1);
             }
         }
-
-        // --- Feed Roller (Triggers) ---
-        // Get trigger values (0.0 to 1.0)
-        float forwardPower = gamepad1.right_trigger;
-        float reversePower = gamepad1.left_trigger;
-
-        // This check is important! It stops the triggers from fighting the 'Y' button command.
-        if (!feedRoller.isBusy()) {
-            // Ensure we are in the correct mode to accept power commands
-            feedRoller.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-            if (forwardPower > 0.1) {
-                // Right trigger pressed: run forward
-                // Your motor direction is REVERSE, so a positive power makes it run "reverse".
-                // If this is the wrong way, just make this -forwardPower
-                feedRoller.setPower(forwardPower);
-            } else if (reversePower > 0.1) {
-                // Left trigger pressed: run in reverse
-                // A negative power will make it run "forward" (opposite of REVERSE)
-                feedRoller.setPower(-reversePower);
-            } else {
-                // No trigger pressed: stop the motor
+        if(gamepad1.xWasPressed()) {
+            if(feedRollerPowered) {
+                feedRollerPowered = false;
                 feedRoller.setPower(0);
+            } else {
+                feedRollerPowered = true;
+                feedRoller.setPower(1);
             }
         }
-
-        // --- All Stop / Reset (Y button) ---
-        if(gamepad1.yWasPressed()) {
-            float angleOff = (feedRoller.getCurrentPosition() % ticksPerRev);
-
-            // feedRollerPowered is no longer needed
-            // feedRollerPowered = false;
-
-            feedRoller.setTargetPosition((int)(feedRoller.getCurrentPosition() - angleOff));
-            feedRoller.setPower(1);
-            feedRoller.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            flywheel.setPower(0);
-            agitator.setPower(0);
+        if(gamepad1.y) {
+            feedRoller.setPower(-0.5);
+        }
+        if(gamepad1.yWasReleased()) {
+            feedRoller.setPower(0);
         }
     }
-
     public void flyWheel() {
         if(flyWheelPowered) {
             double multiplier = 14 / getLowestVoltage();
-            telemetry.addData("Velocity", flyWheelVelocity * multiplier);
             flywheel.setVelocity(flyWheelVelocity * multiplier);
         } else {
             flywheel.setVelocity(0);
