+++ FtcRobotController/src/main/java/org/firstinspires/ftc/robotcontroller/Old/RedAutoWithShooting.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Old;
+package org.firstinspires.ftc.robotcontroller.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -10,15 +10,13 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 @Autonomous
-public class AutoWithMove extends LinearOpMode {
-
+public class RedAutoWithShooting extends LinearOpMode {
     private DcMotorEx flywheel;
     private DcMotor feedRoller;
     private CRServo agitator;
-    private DcMotor leftDrive, rightDrive; // drivetrain motors
-    float ticksPerRev = 28 * 5 * 3; // 5 and 3 is the gear reductions
-    float circumference = 9.5f; // in inches
-    float ticksPerInch = ticksPerRev / circumference;
+    private DcMotor leftDrive;
+    private DcMotor rightDrive;
+
     ElapsedTime time = new ElapsedTime();
     double multiplier;
 
@@ -37,71 +35,46 @@ public class AutoWithMove extends LinearOpMode {
 
     @Override
     public void runOpMode() {
-
+        // Flywheel/shooter hardware
         flywheel = hardwareMap.get(DcMotorEx.class, "flywheel");
         feedRoller = hardwareMap.get(DcMotor.class, "coreHex");
         agitator = hardwareMap.get(CRServo.class, "servo");
 
-        // **Add drivetrain motors**
+        // Drivetrain hardware
         leftDrive = hardwareMap.get(DcMotor.class, "leftDrive");
         rightDrive = hardwareMap.get(DcMotor.class, "rightDrive");
+        rightDrive.setDirection(DcMotorSimple.Direction.REVERSE);
 
         flywheel.setDirection(DcMotorSimple.Direction.REVERSE);
-        feedRoller.setDirection(DcMotorSimple.Direction.FORWARD);
-        leftDrive.setDirection(DcMotorSimple.Direction.FORWARD); // adjust if necessary
-
-        // Reset and run flywheel using encoders
         flywheel.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         flywheel.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        feedRoller.setDirection(DcMotorSimple.Direction.REVERSE);
+
 
         waitForStart();
 
-        // --- Back up 1 foot before starting sequence ---
-        moveBackward(12); // 12 inches = 1 ft
-
         // --- Shooting sequence ---
         multiplier = 14 / getLowestVoltage();
-        ourSleep(1000);
-        agitator.setPower(1);
-        ourSleep(500);
+        ourSleep(1000);           // Spin up flywheel
+        //agitator.setPower(1);
+        //ourSleep(500);            // Agitate
         feedRoller.setPower(1);
-        ourSleep(6000);
+        ourSleep(6000);           // Feed balls
 
-        // Stop everything
+        // Stop shooter motors
         flywheel.setPower(0);
-        agitator.setPower(0);
+        //agitator.setPower(0);
         feedRoller.setPower(0);
-        leftDrive.setPower(0);
-        rightDrive.setPower(0);
-    }
 
-    // --- Helper method to move backward ---
-    void moveBackward(double inches) {
-        int targetTicks = (int)(inches * ticksPerInch);
-
-        leftDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        rightDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
-        leftDrive.setTargetPosition(-targetTicks);  // negative for backward
-        rightDrive.setTargetPosition(-targetTicks);
-
-        leftDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        rightDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-        leftDrive.setPower(0.5);
-        rightDrive.setPower(0.5);
-
-        while (opModeIsActive() && leftDrive.isBusy() && rightDrive.isBusy()) {
-            idle();
-        }
-
-        // Stop motors
+        rightDrive.setPower(-1);
+        leftDrive.setPower(1);
+        sleep(1000);
+        leftDrive.setPower(1);
+        rightDrive.setPower(1);
+        sleep(1000);
         leftDrive.setPower(0);
         rightDrive.setPower(0);
 
-        // Return to RUN_USING_ENCODER for teleop/autonomous control
-        leftDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        rightDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
     }
 
     void ourSleep(double timeTakes) {
@@ -109,10 +82,6 @@ public class AutoWithMove extends LinearOpMode {
         while (time.milliseconds() < timeTakes && opModeIsActive()) {
             multiplier = 14 / getLowestVoltage();
             flywheel.setVelocity(900 * multiplier);
-
-            telemetry.addLine("Multiplier: " + multiplier);
-            telemetry.addLine(String.valueOf(flywheel.getVelocity()));
-            telemetry.update();
             idle();
         }
     }
