+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/GoBuildaTeleop.java
@@ -1,11 +1,13 @@
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
+@TeleOp
 public class GoBuildaTeleop extends OpMode {
     private DcMotor leftFrontDrive = null;
     private DcMotor rightFrontDrive = null;
@@ -13,11 +15,9 @@ public class GoBuildaTeleop extends OpMode {
     private DcMotor rightBackDrive = null;
     private DcMotorEx launcher = null;
     private DcMotorEx intake = null;
-
     private CRServo leftFeeder = null;
     private CRServo rightFeeder = null;
     final double targetVelocity = 1125;
-    final double minVelocity = 1075;
 
     boolean launcherOn = false;
     boolean intakeOn = false;
@@ -29,6 +29,8 @@ public class GoBuildaTeleop extends OpMode {
         rightBackDrive = hardwareMap.get(DcMotor.class, "br");
         launcher = hardwareMap.get(DcMotorEx.class, "launcher");
         intake = hardwareMap.get(DcMotorEx.class, "intake");
+        leftFeeder = hardwareMap.get(CRServo.class, "left_feeder");
+        rightFeeder = hardwareMap.get(CRServo.class, "right_feeder");
 
         leftFrontDrive.setDirection(DcMotor.Direction.REVERSE);
         rightFrontDrive.setDirection(DcMotor.Direction.FORWARD);
@@ -38,32 +40,52 @@ public class GoBuildaTeleop extends OpMode {
 
         launcher.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
+        leftFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        rightFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        leftBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        rightBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
         telemetry.addLine("Y to turn launcher off/on");
         telemetry.addLine("A to turn intake off/on");
         telemetry.addLine("X to turn reverse intake direction");
+        telemetry.addLine("Right & Left trigger for feeders");
     }
     public void loop() {
         drive();
         turnOn();
     }
     public void turnOn() {
-        if (gamepad1.y) {
+        if (gamepad1.yWasReleased()) {
             launcherOn = !launcherOn;
             launcher.setVelocity(launcherOn ? targetVelocity : 0);
-        } else if(gamepad1.a){
+        }
+        if(gamepad1.aWasReleased()){
             intakeOn = !intakeOn;
             intake.setPower(intakeOn ? 1 : 0);
-        } else if(gamepad1.x){
+        }
+        if(gamepad1.xWasReleased()){
             intakeIn = !intakeIn;
-            intake.setDirection(intakeIn ? DcMotorSimple.Direction.FORWARD : DcMotorSimple.Direction.REVERSE);
+            intake.setDirection(intakeIn ? DcMotorEx.Direction.FORWARD : DcMotorEx.Direction.REVERSE);
+        }
+
+        if(gamepad1.left_trigger > 0.1) {
+            leftFeeder.setPower(1);
+            rightFeeder.setPower(1);
+        }
+        else if(gamepad1.right_trigger > 0.1) {
+            leftFeeder.setPower(-1);
+            rightFeeder.setPower(-1);
+        } else {
+            leftFeeder.setPower(0);
+            rightFeeder.setPower(0);
         }
     }
     public void drive() {
-        double denominator = Math.max(Math.abs(gamepad1.left_stick_y) + Math.abs(gamepad1.left_stick_x) + Math.abs(gamepad1.right_stick_x), 1);
+        double denominator = Math.max(Math.abs(-gamepad1.left_stick_y) + Math.abs(gamepad1.left_stick_x) + Math.abs(gamepad1.right_stick_x), 1);
 
-        leftFrontDrive.setPower((gamepad1.left_stick_y + gamepad1.left_stick_x + gamepad1.right_stick_x) / denominator);
-        rightFrontDrive.setPower((gamepad1.left_stick_y - gamepad1.left_stick_x - gamepad1.right_stick_x) / denominator);
-        leftBackDrive.setPower((gamepad1.left_stick_y - gamepad1.left_stick_x + gamepad1.right_stick_x) / denominator);
-        rightBackDrive.setPower((gamepad1.left_stick_y + gamepad1.left_stick_x - gamepad1.right_stick_x) / denominator);
+        leftFrontDrive.setPower((-gamepad1.left_stick_y + gamepad1.left_stick_x + gamepad1.right_stick_x) / denominator);
+        rightFrontDrive.setPower((-gamepad1.left_stick_y - gamepad1.left_stick_x - gamepad1.right_stick_x) / denominator);
+        leftBackDrive.setPower((-gamepad1.left_stick_y - gamepad1.left_stick_x + gamepad1.right_stick_x) / denominator);
+        rightBackDrive.setPower((-gamepad1.left_stick_y + gamepad1.left_stick_x - gamepad1.right_stick_x) / denominator);
     }
 }
